Complete full CRUD for portfolio

+++ src/main/java/com/example/investmentrestapi/repository/PortfolioRepository.java
@@ -4,6 +4,6 @@ package com.example.investmentrestapi.repository;
 import com.example.investmentrestapi.model.Portfolio;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PortfolioRepository extends CrudRepository<Portfolio, Long> {
+public interface PortfolioRepository extends CrudRepository<Portfolio, Integer> {
 
 }
